backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1

pydantic==2.10.2
pydantic-settings==2.6.1

sqlalchemy==2.0.36
psycopg[binary]==3.2.3

python-multipart==0.0.17
pdfplumber==0.11.4
python-docx==1.1.2

langchain==0.3.8
langchain-openai==0.2.9
langchain-anthropic==0.3.0
langchain-google-genai==2.0.8
google-genai==1.3.0

tenacity==9.0.0
python-dotenv==1.0.1

pytest==8.3.3
httpx==0.28.1
ruff==0.7.4
black==24.10.0

pydantic-settings==2.6.1
reportlab==4.0.8